requirements.txt
ollama
pydantic
python-dotenv
